feat: break everything up into seperate files
