+++ src/library/scan.rs
@@ -1,43 +1,40 @@
-#![allow(clippy::explicit_auto_deref)]
+mod database;
+mod decode;
+mod discover;
+mod record;
 
 use std::{
-    io::{Cursor, ErrorKind},
-    path::Path,
     sync::Arc,
     time::{Duration, SystemTime, UNIX_EPOCH},
 };
 
-use async_compression::tokio::write::{ZlibDecoder, ZlibEncoder};
-use camino::{Utf8Path, Utf8PathBuf};
-use globwalk::GlobWalkerBuilder;
+use camino::Utf8PathBuf;
 use gpui::{App, Global};
-use image::{DynamicImage, EncodableLayout, codecs::jpeg::JpegEncoder, imageops};
 
 use rustc_hash::{FxHashMap, FxHashSet};
-use serde::{Deserialize, Serialize};
-use sqlx::{SqliteConnection, SqlitePool};
+use sqlx::SqlitePool;
 use tokio::{
-    io::{AsyncReadExt, AsyncWriteExt},
     sync::{
         Mutex,
         mpsc::{Receiver, Sender, UnboundedReceiver, UnboundedSender, channel, unbounded_channel},
     },
     task::spawn_blocking,
 };
-use tracing::{debug, error, info, warn};
-
-/// The version of the scanning process. If this version number is incremented, a re-scan of all
-/// files will be forced (see [ScanCommand::ForceScan]).
-const SCAN_VERSION: u16 = 1;
-
-/// Maximum number of items to accumulate before flushing a DB transaction.
-const BATCH_SIZE: usize = 50;
+use tracing::{error, info, warn};
 
 use crate::{
-    media::{builtin::symphonia::SymphoniaProvider, metadata::Metadata, traits::MediaProvider},
+    library::scan::{
+        database::{AlbumCacheKey, AlbumPathCacheKey, update_metadata},
+        decode::{FileInformation, build_provider_table, read_metadata_for_path},
+        discover::{cleanup, cleanup_removed_directories, discover},
+        record::{SCAN_VERSION, ScanRecord, load_scan_record, write_scan_record},
+    },
     settings::scan::ScanSettings,
     ui::{app::get_dirs, models::Models},
 };
+
+/// Maximum number of items to accumulate before flushing a DB transaction.
+const BATCH_SIZE: usize = 50;
 
 #[derive(Debug, PartialEq, Clone, Copy)]
 pub enum ScanEvent {
@@ -120,702 +117,6 @@ impl ScanInterface {
 }
 
 impl Global for ScanInterface {}
-
-/// Information extracted from a media file during the metadata reading stage.
-/// Raw image bytes are passed through the pipeline; image processing (resize + thumbnail) only
-/// happens in `insert_album` when a new album is actually created.
-type FileInformation = (Metadata, u64, Option<Box<[u8]>>);
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct ScanRecord {
-    version: u16,
-    records: FxHashMap<Utf8PathBuf, SystemTime>,
-    directories: Vec<Utf8PathBuf>,
-}
-
-impl ScanRecord {
-    fn new_current() -> Self {
-        Self {
-            version: SCAN_VERSION,
-            records: FxHashMap::default(),
-            directories: Vec::new(),
-        }
-    }
-
-    fn is_version_mismatch(&self) -> bool {
-        self.version != SCAN_VERSION
-    }
-}
-
-fn build_provider_table() -> Vec<(Vec<String>, Box<dyn MediaProvider>)> {
-    // TODO: dynamic plugin loading
-    let provider = SymphoniaProvider;
-    vec![(
-        provider
-            .supported_extensions()
-            .iter()
-            .copied()
-            .map(str::to_string)
-            .collect(),
-        Box::new(provider),
-    )]
-}
-
-fn file_is_scannable_with_provider(path: &Utf8Path, exts: &[String]) -> bool {
-    for extension in exts.iter() {
-        if let Some(ext) = path.extension()
-            && *ext == **extension
-        {
-            return true;
-        }
-    }
-
-    false
-}
-
-/// Read metadata, duration, and embedded image from a file using the given provider.
-/// Returns raw (unprocessed) image bytes.
-fn scan_file_with_provider(
-    path: &Utf8Path,
-    provider: &mut Box<dyn MediaProvider>,
-) -> Result<FileInformation, ()> {
-    let src = std::fs::File::open(path).map_err(|_| ())?;
-    let mut stream = provider.open(src, None).map_err(|_| ())?;
-    stream.start_playback().map_err(|_| ())?;
-    let metadata = stream.read_metadata().cloned().map_err(|_| ())?;
-    let image = stream.read_image().map_err(|_| ())?;
-    let len = stream.duration_secs().map_err(|_| ())?;
-    stream.close().map_err(|_| ())?;
-    Ok((metadata, len, image))
-}
-
-/// Returns the first image (cover/front/folder.jpeg/png/jpg) in the track's containing folder.
-/// Results are cached per-directory in `art_cache` to avoid redundant glob walks when multiple
-/// tracks share the same folder.
-fn scan_path_for_album_art(
-    path: &Utf8Path,
-    art_cache: &mut FxHashMap<Utf8PathBuf, Option<Arc<[u8]>>>,
-) -> Option<Arc<[u8]>> {
-    let parent = path.parent()?.to_path_buf();
-
-    if let Some(cached) = art_cache.get(&parent) {
-        return cached.clone();
-    }
-
-    let glob = GlobWalkerBuilder::from_patterns(&parent, &["{folder,cover,front}.{jpg,jpeg,png}"])
-        .case_insensitive(true)
-        .max_depth(1)
-        .build()
-        .expect("Failed to build album art glob")
-        .filter_map(|e| e.ok());
-
-    for entry in glob {
-        if let Ok(bytes) = std::fs::read(entry.path()) {
-            let arc: Arc<[u8]> = Arc::from(bytes);
-            art_cache.insert(parent, Some(Arc::clone(&arc)));
-            return Some(arc);
-        }
-    }
-
-    art_cache.insert(parent, None);
-    None
-}
-
-/// Check if a file should be scanned.
-/// Returns `Some(timestamp)` if the file should be scanned (not in scan_record or modified since last scan).
-/// Returns `None` if the file should be skipped or cannot be scanned.
-fn file_is_scannable(
-    path: &Utf8Path,
-    scan_record: &FxHashMap<Utf8PathBuf, SystemTime>,
-    provider_table: &[(Vec<String>, Box<dyn MediaProvider>)],
-) -> Option<SystemTime> {
-    let Ok(timestamp) = (match std::fs::metadata(path) {
-        Ok(metadata) => metadata.modified(),
-        Err(_) => return None,
-    }) else {
-        return None;
-    };
-
-    for (exts, _) in provider_table.iter() {
-        let x = file_is_scannable_with_provider(path, exts);
-
-        if !x {
-            continue;
-        }
-
-        if let Some(last_scan) = scan_record.get(path)
-            && *last_scan == timestamp
-        {
-            return None;
-        }
-
-        return Some(timestamp);
-    }
-
-    None
-}
-
-/// Process album art into a (resized_full_image, thumbnail_bmp) pair.
-///
-/// The thumbnail is always a 70×70 BMP. The full-size image is passed through if both dimensions
-/// are ≤ 1024, otherwise it is downscaled to 1024×1024 and re-encoded as JPEG.
-fn process_album_art(image: &[u8]) -> anyhow::Result<(Vec<u8>, Vec<u8>)> {
-    let decoded = image::ImageReader::new(Cursor::new(image))
-        .with_guessed_format()?
-        .decode()?
-        .into_rgb8();
-
-    // thumbnail
-    let thumb_rgb = imageops::thumbnail(&decoded, 70, 70);
-    let thumb_rgba = DynamicImage::ImageRgb8(thumb_rgb).into_rgba8();
-
-    let mut thumb_buf: Vec<u8> = Vec::new();
-    thumb_rgba.write_to(&mut Cursor::new(&mut thumb_buf), image::ImageFormat::Bmp)?;
-
-    // full-size image (resized if necessary)
-    let resized = if decoded.dimensions().0 <= 1024 && decoded.dimensions().1 <= 1024 {
-        image.to_vec()
-    } else {
-        // preserve aspect ratio
-        let (w, h) = decoded.dimensions();
-        let scale = 1024.0_f32 / (w.max(h) as f32);
-        let new_w = (w as f32 * scale).round().max(1.0) as u32;
-        let new_h = (h as f32 * scale).round().max(1.0) as u32;
-
-        let resized_img = imageops::resize(
-            &decoded,
-            new_w,
-            new_h,
-            image::imageops::FilterType::Lanczos3,
-        );
-        let mut buf: Cursor<Vec<u8>> = Cursor::new(Vec::new());
-        let mut encoder = JpegEncoder::new_with_quality(&mut buf, 70);
-
-        encoder.encode(
-            resized_img.as_bytes(),
-            resized_img.width(),
-            resized_img.height(),
-            image::ExtendedColorType::Rgb8,
-        )?;
-        drop(encoder);
-
-        buf.into_inner()
-    };
-
-    Ok((resized, thumb_buf))
-}
-
-/// Read metadata from a file, resolve album art (embedded or from directory).
-///
-/// Each metadata reader thread maintains its own `art_cache` to avoid redundant directory scans
-/// for files in the same folder.
-fn read_metadata_for_path(
-    path: &Utf8Path,
-    provider_table: &mut Vec<(Vec<String>, Box<dyn MediaProvider>)>,
-    art_cache: &mut FxHashMap<Utf8PathBuf, Option<Arc<[u8]>>>,
-) -> Option<FileInformation> {
-    for (exts, provider) in provider_table.iter_mut() {
-        if file_is_scannable_with_provider(path, exts)
-            && let Ok(mut metadata) = scan_file_with_provider(path, provider)
-        {
-            if metadata.2.is_none()
-                && let Some(art) = scan_path_for_album_art(path, art_cache)
-            {
-                metadata.2 = Some(art.to_vec().into_boxed_slice());
-            }
-
-            return Some(metadata);
-        }
-    }
-
-    None
-}
-
-async fn load_scan_record(path: &Path) -> ScanRecord {
-    let mut file = match tokio::fs::File::open(path).await.map(ZlibDecoder::new) {
-        Ok(f) => f,
-        Err(e) => {
-            if e.kind() != ErrorKind::NotFound {
-                error!("Could not open scan record: {:?}", e);
-                error!("Scanning will be slow until the scan record is rebuilt");
-            }
-
-            return ScanRecord::new_current();
-        }
-    };
-
-    let mut bytes = Vec::new();
-    file.read_to_end(&mut bytes).await.unwrap_or_default();
-
-    match postcard::from_bytes(&bytes) {
-        Ok(scan_record) => scan_record,
-        Err(e) => {
-            error!("Could not read scan record: {:?}", e);
-            error!("Scanning will be slow until the scan record is rebuilt");
-            ScanRecord::new_current()
-        }
-    }
-}
-
-async fn write_scan_record(scan_record: &ScanRecord, path: &Path) {
-    let mut file = match tokio::fs::File::create(path).await.map(ZlibEncoder::new) {
-        Ok(file) => file,
-        Err(e) => {
-            error!("Could not create scan record file: {:?}", e);
-            error!("Scan record will not be saved, this may cause rescans on restart");
-            return;
-        }
-    };
-
-    match postcard::to_allocvec(&scan_record) {
-        Ok(data) => match file.write_all(&data).await {
-            Ok(_) => {
-                info!("Scan record saved successfully");
-            }
-            Err(e) => {
-                error!("Could not write scan record: {:?}", e);
-                error!("Scan record will not be saved, this may cause rescans on restart");
-            }
-        },
-        Err(e) => {
-            error!("Could not serialize scan record: {:?}", e);
-            error!("Scan record will not be saved, this may cause rescans on restart");
-        }
-    }
-}
-
-/// Performs a full recursive directory walk, streaming discovered file paths through `path_tx`
-/// as they are found so that downstream pipeline stages can begin processing immediately.
-///
-/// Returns the total number of discovered files once the walk is complete.
-fn discover(
-    settings: ScanSettings,
-    scan_record: Arc<Mutex<ScanRecord>>,
-    path_tx: Sender<(Utf8PathBuf, SystemTime)>,
-) -> u64 {
-    let provider_table = build_provider_table();
-    let mut visited: FxHashSet<Utf8PathBuf> = FxHashSet::default();
-    let mut stack: Vec<Utf8PathBuf> = settings.paths.clone();
-    let mut discovered_total: u64 = 0;
-
-    while let Some(dir) = stack.pop() {
-        if !visited.insert(dir.clone()) {
-            continue;
-        }
-
-        let entries = match std::fs::read_dir(&dir) {
-            Ok(e) => e,
-            Err(e) => {
-                error!("Failed to read directory {:?}: {:?}", dir, e);
-                continue;
-            }
-        };
-
-        for entry in entries {
-            let path = match entry {
-                Ok(entry) => match entry.path().canonicalize() {
-                    Ok(p) => match Utf8PathBuf::try_from(p) {
-                        Ok(u) => u,
-                        Err(e) => {
-                            error!(
-                                "Failed to convert path {:?} to UTF-8: {:?}",
-                                entry.path(),
-                                e
-                            );
-                            continue;
-                        }
-                    },
-                    Err(e) => {
-                        error!("Failed to canonicalize path {:?}: {:?}", entry.path(), e);
-                        continue;
-                    }
-                },
-                Err(e) => {
-                    error!("Failed to read directory entry: {:?}", e);
-                    continue;
-                }
-            };
-
-            if path.is_dir() {
-                stack.push(path);
-            } else {
-                let timestamp = {
-                    let sr = scan_record.blocking_lock();
-                    file_is_scannable(&path, &sr.records, &provider_table)
-                };
-
-                if let Some(ts) = timestamp {
-                    discovered_total += 1;
-
-                    if path_tx.blocking_send((path, ts)).is_err() {
-                        return discovered_total;
-                    }
-                }
-            }
-        }
-    }
-
-    discovered_total
-}
-
-async fn insert_artist(
-    conn: &mut SqliteConnection,
-    metadata: &Metadata,
-    artist_cache: &mut FxHashMap<String, i64>,
-) -> anyhow::Result<Option<i64>> {
-    let artist = metadata.album_artist.clone().or(metadata.artist.clone());
-
-    let Some(artist) = artist else {
-        return Ok(None);
-    };
-
-    // Check in-memory cache first
-    if let Some(&cached_id) = artist_cache.get(&artist) {
-        return Ok(Some(cached_id));
-    }
-
-    let result: Result<(i64,), sqlx::Error> =
-        sqlx::query_as(include_str!("../../queries/scan/create_artist.sql"))
-            .bind(&artist)
-            .bind(metadata.artist_sort.as_ref().unwrap_or(&artist))
-            .fetch_one(&mut *conn)
-            .await;
-
-    let id = match result {
-        Ok(v) => v.0,
-        Err(sqlx::Error::RowNotFound) => {
-            let result: Result<(i64,), sqlx::Error> =
-                sqlx::query_as(include_str!("../../queries/scan/get_artist_id.sql"))
-                    .bind(&artist)
-                    .fetch_one(&mut *conn)
-                    .await;
-
-            match result {
-                Ok(v) => v.0,
-                Err(e) => return Err(e.into()),
-            }
-        }
-        Err(e) => return Err(e.into()),
-    };
-
-    artist_cache.insert(artist, id);
-    Ok(Some(id))
-}
-
-/// Album cache key: (title, mbid, artist_id).
-type AlbumCacheKey = (String, String, Option<i64>);
-
-async fn insert_album(
-    conn: &mut SqliteConnection,
-    metadata: &Metadata,
-    artist_id: Option<i64>,
-    image: &Option<Box<[u8]>>,
-    is_force: bool,
-    force_encountered_albums: &mut FxHashSet<i64>,
-    album_cache: &mut FxHashMap<AlbumCacheKey, i64>,
-) -> anyhow::Result<Option<i64>> {
-    let Some(album) = &metadata.album else {
-        return Ok(None);
-    };
-
-    let mbid = metadata
-        .mbid_album
-        .clone()
-        .unwrap_or_else(|| "none".to_string());
-
-    let cache_key: AlbumCacheKey = (album.clone(), mbid.clone(), artist_id);
-
-    if !is_force && let Some(&cached_id) = album_cache.get(&cache_key) {
-        return Ok(Some(cached_id));
-    }
-
-    let result: Result<(i64,), sqlx::Error> =
-        sqlx::query_as(include_str!("../../queries/scan/get_album_id.sql"))
-            .bind(album)
-            .bind(&mbid)
-            .bind(artist_id)
-            .fetch_one(&mut *conn)
-            .await;
-
-    let should_force = if let Ok((id,)) = &result
-        && is_force
-    {
-        force_encountered_albums.insert(*id)
-    } else {
-        false
-    };
-
-    match (result, should_force) {
-        (Ok(v), false) => {
-            album_cache.insert(cache_key, v.0);
-            Ok(Some(v.0))
-        }
-        (Err(sqlx::Error::RowNotFound), _) | (Ok(_), true) => {
-            let (resized_image, thumb) = match image {
-                Some(image) => {
-                    match process_album_art(image) {
-                        Ok((resized, thumb)) => (Some(resized), Some(thumb)),
-                        Err(e) => {
-                            // if there is a decode error, just ignore it and pretend there is no image
-                            warn!("Failed to process album art: {:?}", e);
-                            (None, None)
-                        }
-                    }
-                }
-                None => (None, None),
-            };
-
-            let result: (i64,) =
-                sqlx::query_as(include_str!("../../queries/scan/create_album.sql"))
-                    .bind(album)
-                    .bind(metadata.sort_album.as_ref().unwrap_or(album))
-                    .bind(artist_id)
-                    .bind(resized_image.as_deref())
-                    .bind(thumb.as_deref())
-                    .bind(metadata.date)
-                    .bind(metadata.year)
-                    .bind(&metadata.label)
-                    .bind(&metadata.catalog)
-                    .bind(&metadata.isrc)
-                    .bind(&mbid)
-                    .bind(metadata.vinyl_numbering)
-                    .fetch_one(&mut *conn)
-                    .await?;
-
-            album_cache.insert(cache_key, result.0);
-            Ok(Some(result.0))
-        }
-        (Err(e), _) => Err(e.into()),
-    }
-}
-
-/// Album-path cache key: (album_id, disc_num).
-type AlbumPathCacheKey = (i64, i64);
-
-async fn insert_track(
-    conn: &mut SqliteConnection,
-    metadata: &Metadata,
-    album_id: Option<i64>,
-    path: &Utf8Path,
-    length: u64,
-    album_path_cache: &mut FxHashMap<AlbumPathCacheKey, Utf8PathBuf>,
-) -> anyhow::Result<()> {
-    if album_id.is_none() {
-        return Ok(());
-    }
-
-    let album_id_val = album_id.unwrap();
-    let disc_num = metadata.disc_current.map(|v| v as i64).unwrap_or(-1);
-    let parent = path.parent().unwrap();
-    let ap_key = (album_id_val, disc_num);
-
-    // Check album-path cache first to avoid DB round-trips
-    if let Some(cached_path) = album_path_cache.get(&ap_key) {
-        if cached_path.as_path() != parent {
-            return Ok(());
-        }
-    } else {
-        let find_path: Result<(String,), _> =
-            sqlx::query_as(include_str!("../../queries/scan/get_album_path.sql"))
-                .bind(album_id)
-                .bind(disc_num)
-                .fetch_one(&mut *conn)
-                .await;
-
-        match find_path {
-            Ok(found) => {
-                let found_path = Utf8PathBuf::from(&found.0);
-                album_path_cache.insert(ap_key, found_path.clone());
-                if found_path.as_path() != parent {
-                    return Ok(());
-                }
-            }
-            Err(sqlx::Error::RowNotFound) => {
-                sqlx::query(include_str!("../../queries/scan/create_album_path.sql"))
-                    .bind(album_id)
-                    .bind(parent.as_str())
-                    .bind(disc_num)
-                    .execute(&mut *conn)
-                    .await?;
-                album_path_cache.insert(ap_key, parent.to_path_buf());
-            }
-            Err(e) => return Err(e.into()),
-        }
-    }
-
-    let name = metadata
-        .name
-        .clone()
-        .or_else(|| path.file_name().map(|v| v.to_string()))
-        .ok_or_else(|| anyhow::anyhow!("failed to retrieve filename"))?;
-
-    let result: Result<(i64,), sqlx::Error> =
-        sqlx::query_as(include_str!("../../queries/scan/create_track.sql"))
-            .bind(&name)
-            .bind(&name)
-            .bind(album_id)
-            .bind(metadata.track_current.map(|x| x as i32))
-            .bind(metadata.disc_current.map(|x| x as i32))
-            .bind(length as i32)
-            .bind(path.as_str())
-            .bind(&metadata.genre)
-            .bind(&metadata.artist)
-            .bind(parent.as_str())
-            .fetch_one(&mut *conn)
-            .await;
-
-    match result {
-        Ok(_) => Ok(()),
-        Err(sqlx::Error::RowNotFound) => Ok(()),
-        Err(e) => Err(e.into()),
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn update_metadata(
-    conn: &mut SqliteConnection,
-    metadata: &Metadata,
-    path: &Utf8Path,
-    length: u64,
-    image: &Option<Box<[u8]>>,
-    is_force: bool,
-    force_encountered_albums: &mut FxHashSet<i64>,
-    artist_cache: &mut FxHashMap<String, i64>,
-    album_cache: &mut FxHashMap<AlbumCacheKey, i64>,
-    album_path_cache: &mut FxHashMap<AlbumPathCacheKey, Utf8PathBuf>,
-) -> anyhow::Result<()> {
-    debug!(
-        "Adding/updating record for {:?} - {:?}",
-        metadata.artist, metadata.name
-    );
-
-    let artist_id = insert_artist(conn, metadata, artist_cache).await?;
-    let album_id = insert_album(
-        conn,
-        metadata,
-        artist_id,
-        image,
-        is_force,
-        force_encountered_albums,
-        album_cache,
-    )
-    .await?;
-    insert_track(conn, metadata, album_id, path, length, album_path_cache).await?;
-
-    Ok(())
-}
-
-/// Remove tracks from directories that are no longer in the scan configuration.
-async fn cleanup_removed_directories(
-    pool: &SqlitePool,
-    scan_record: &mut ScanRecord,
-    current_directories: &[Utf8PathBuf],
-) {
-    let current_set: FxHashSet<Utf8PathBuf> = current_directories.iter().cloned().collect();
-    let old_set: FxHashSet<Utf8PathBuf> = scan_record.directories.iter().cloned().collect();
-
-    let removed_dirs: Vec<Utf8PathBuf> = old_set.difference(&current_set).cloned().collect();
-
-    if removed_dirs.is_empty() {
-        return;
-    }
-
-    info!(
-        "Detected {} removed directories, cleaning up tracks",
-        removed_dirs.len()
-    );
-
-    let mut tx = match pool.begin().await {
-        Ok(tx) => tx,
-        Err(e) => {
-            error!("Could not begin directory cleanup transaction: {:?}", e);
-            return;
-        }
-    };
-
-    let to_remove: Vec<Utf8PathBuf> = scan_record
-        .records
-        .keys()
-        .filter(|path| {
-            removed_dirs
-                .iter()
-                .any(|removed_dir| path.starts_with(removed_dir))
-        })
-        .cloned()
-        .collect();
-
-    let mut deleted: Vec<Utf8PathBuf> = Vec::with_capacity(to_remove.len());
-    for path in &to_remove {
-        debug!("removing track from removed directory: {:?}", path);
-        let result = sqlx::query(include_str!("../../queries/scan/delete_track.sql"))
-            .bind(path.as_str())
-            .execute(&mut *tx)
-            .await;
-
-        if let Err(e) = result {
-            error!("Database error while deleting track: {:?}", e);
-        } else {
-            deleted.push(path.clone());
-        }
-    }
-
-    if let Err(e) = tx.commit().await {
-        error!("Failed to commit directory cleanup transaction: {:?}", e);
-        return;
-    }
-
-    for path in &deleted {
-        scan_record.records.remove(path);
-    }
-
-    info!(
-        "Cleaned up {} track(s) from removed directories",
-        deleted.len()
-    );
-}
-
-/// Remove scan_record entries whose files no longer exist on disk, and delete the corresponding
-/// tracks from the database.
-async fn cleanup(pool: &SqlitePool, scan_record: &mut ScanRecord) {
-    let to_delete: Vec<Utf8PathBuf> = scan_record
-        .records
-        .keys()
-        .filter(|path| !path.exists())
-        .cloned()
-        .collect();
-
-    let mut tx = match pool.begin().await {
-        Ok(tx) => tx,
-        Err(e) => {
-            error!("Could not begin cleanup transaction: {:?}", e);
-            return;
-        }
-    };
-
-    let mut deleted: Vec<Utf8PathBuf> = Vec::with_capacity(to_delete.len());
-    for path in &to_delete {
-        debug!("track deleted or moved: {:?}", path);
-        let result = sqlx::query(include_str!("../../queries/scan/delete_track.sql"))
-            .bind(path.as_str())
-            .execute(&mut *tx)
-            .await;
-
-        if let Err(e) = result {
-            error!("Database error while deleting track: {:?}", e);
-        } else {
-            deleted.push(path.clone());
-        }
-    }
-
-    if let Err(e) = tx.commit().await {
-        error!("Failed to commit cleanup transaction: {:?}", e);
-        return;
-    }
-
-    for path in &deleted {
-        scan_record.records.remove(path);
-    }
-}
 
 async fn run_scanner(
     pool: SqlitePool,
